+++ src/sam/writer.jl
@@ -10,7 +10,7 @@ Create a data writer of the SAM file format.
 * `output`: data sink
 * `header=Header()`: SAM header object
 """
-mutable struct Writer <: BioGenerics.IO.AbstractWriter
+mutable struct Writer <: XAMWriter
     stream::IO
 
     function Writer(output::IO, header::Header=Header())