+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added BAM.Reader index support for BAI object ([#56](https://github.com/BioJulia/XAM.jl/pull/56/files))
 
+### Changed
+
+- Subtype from XAMReader and XAMWriter from common abstract types.
+
 ## [0.3.1]
 
 ### Changed